+++ etna/experimental/prediction_intervals/__init__.py
@@ -1,2 +1,3 @@
 from etna.experimental.prediction_intervals.base import BasePredictionIntervals
+from etna.experimental.prediction_intervals.conformal import ConformalPredictionIntervals
 from etna.experimental.prediction_intervals.naive_variance import NaiveVariancePredictionIntervals