+++ tests/test_experimental/test_prediction_intervals/test_conformal.py
@@ -7,7 +7,7 @@ import pytest
 from etna.ensembles import DirectEnsemble
 from etna.ensembles import StackingEnsemble
 from etna.ensembles import VotingEnsemble
-from etna.experimental.prediction_intervals import NaiveVariancePredictionIntervals
+from etna.experimental.prediction_intervals import ConformalPredictionIntervals
 from etna.models import NaiveModel
 from etna.pipeline import AutoRegressivePipeline
 from etna.pipeline import HierarchicalPipeline
@@ -20,23 +20,23 @@ from tests.test_experimental.test_prediction_intervals.common import get_naive_p
 from tests.test_experimental.test_prediction_intervals.common import run_base_pipeline_compat_check
 
 
-def test_invalid_stride_error():
+@pytest.mark.parametrize("stride", (-1, 0))
+def test_invalid_stride_parameter_error(stride):
     with pytest.raises(ValueError, match="Parameter `stride` must be positive!"):
-        NaiveVariancePredictionIntervals(pipeline=Pipeline(model=NaiveModel()), stride=-1)
+        ConformalPredictionIntervals(pipeline=Pipeline(model=NaiveModel()), stride=stride)
 
 
-@pytest.mark.parametrize("dummy_array", (np.ones(shape=(3, 5, 3)),))
-def test_estimate_variance_shape(dummy_array):
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=Pipeline(model=NaiveModel()))
-    out_array = intervals_pipeline._estimate_variance(residual_matrices=dummy_array)
-    assert out_array.shape == (5, 3)
+@pytest.mark.parametrize("coverage", (-3, -1))
+def test_invalid_coverage_parameter_error(coverage):
+    with pytest.raises(ValueError, match="Parameter `coverage` must be non-negative"):
+        ConformalPredictionIntervals(pipeline=Pipeline(model=NaiveModel()), coverage=coverage)
 
 
 @pytest.mark.parametrize("pipeline_name", ("naive_pipeline", "naive_pipeline_with_transforms"))
 def test_pipeline_fit_forecast_without_intervals(example_tsds, pipeline_name, request):
     pipeline = request.getfixturevalue(pipeline_name)
 
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=pipeline)
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=pipeline)
 
     intervals_pipeline.fit(ts=example_tsds)
 
@@ -47,11 +47,9 @@ def test_pipeline_fit_forecast_without_intervals(example_tsds, pipeline_name, re
 
 
 @pytest.mark.parametrize("stride", (2, 5, 10))
-@pytest.mark.parametrize("expected_columns", ({"target", "target_0.025", "target_0.975"},))
+@pytest.mark.parametrize("expected_columns", ({"target", "target_lower", "target_upper"},))
 def test_valid_strides(example_tsds, expected_columns, stride):
-    intervals_pipeline = NaiveVariancePredictionIntervals(
-        pipeline=Pipeline(model=NaiveModel(), horizon=5), stride=stride
-    )
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=Pipeline(model=NaiveModel(), horizon=5), stride=stride)
     run_base_pipeline_compat_check(
         ts=example_tsds, intervals_pipeline=intervals_pipeline, expected_columns=expected_columns
     )
@@ -59,7 +57,7 @@ def test_valid_strides(example_tsds, expected_columns, stride):
 
 @pytest.mark.parametrize(
     "expected_columns",
-    ({"target", "target_0.025", "target_0.975"},),
+    ({"target", "target_lower", "target_upper"},),
 )
 @pytest.mark.parametrize(
     "pipeline",
@@ -75,7 +73,7 @@ def test_valid_strides(example_tsds, expected_columns, stride):
     ),
 )
 def test_pipelines_forecast_intervals_exist(product_level_constant_hierarchical_ts, pipeline, expected_columns):
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=pipeline)
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=pipeline)
     run_base_pipeline_compat_check(
         ts=product_level_constant_hierarchical_ts,
         intervals_pipeline=intervals_pipeline,
@@ -85,7 +83,7 @@ def test_pipelines_forecast_intervals_exist(product_level_constant_hierarchical_
 
 @pytest.mark.parametrize("pipeline", (get_arima_pipeline(horizon=5),))
 def test_forecast_prediction_intervals_is_used(example_tsds, pipeline):
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=pipeline)
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=pipeline)
     intervals_pipeline._forecast_prediction_interval = MagicMock()
 
     intervals_pipeline.fit(ts=example_tsds)
@@ -104,18 +102,17 @@ def test_forecast_prediction_intervals_is_used(example_tsds, pipeline):
     ),
 )
 def test_pipelines_forecast_intervals_valid(example_tsds, pipeline):
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=pipeline)
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=pipeline)
     intervals_pipeline.fit(ts=example_tsds)
 
-    prediction = intervals_pipeline.forecast(prediction_interval=True, quantiles=(0.025, 0.1, 0.975))
-    assert np.all(prediction[:, :, "target_0.025"].values <= prediction[:, :, "target_0.1"].values)
-    assert np.all(prediction[:, :, "target_0.1"].values <= prediction[:, :, "target"].values)
-    assert np.all(prediction[:, :, "target"].values <= prediction[:, :, "target_0.975"].values)
+    prediction = intervals_pipeline.forecast(prediction_interval=True)
+    assert np.all(prediction[:, :, "target_lower"].values <= prediction[:, :, "target"].values)
+    assert np.all(prediction[:, :, "target"].values <= prediction[:, :, "target_upper"].values)
 
 
 @pytest.mark.parametrize(
     "expected_columns",
-    ({"target", "target_0.025", "target_0.975"},),
+    ({"target", "target_lower", "target_upper"},),
 )
 @pytest.mark.parametrize(
     "ensemble",
@@ -126,7 +123,7 @@ def test_pipelines_forecast_intervals_valid(example_tsds, pipeline):
     ),
 )
 def test_ensembles_forecast_intervals_exist(example_tsds, ensemble, expected_columns):
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=ensemble)
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=ensemble)
     run_base_pipeline_compat_check(
         ts=example_tsds, intervals_pipeline=intervals_pipeline, expected_columns=expected_columns
     )
@@ -141,10 +138,9 @@ def test_ensembles_forecast_intervals_exist(example_tsds, ensemble, expected_col
     ),
 )
 def test_ensembles_forecast_intervals_valid(example_tsds, ensemble):
-    intervals_pipeline = NaiveVariancePredictionIntervals(pipeline=ensemble)
+    intervals_pipeline = ConformalPredictionIntervals(pipeline=ensemble)
     intervals_pipeline.fit(ts=example_tsds)
 
-    prediction = intervals_pipeline.forecast(prediction_interval=True, quantiles=(0.025, 0.1, 0.975))
-    assert np.all(prediction[:, :, "target_0.025"].values <= prediction[:, :, "target_0.1"].values)
-    assert np.all(prediction[:, :, "target_0.1"].values <= prediction[:, :, "target"].values)
-    assert np.all(prediction[:, :, "target"].values <= prediction[:, :, "target_0.975"].values)
+    prediction = intervals_pipeline.forecast(prediction_interval=True)
+    assert np.all(prediction[:, :, "target_lower"].values <= prediction[:, :, "target"].values)
+    assert np.all(prediction[:, :, "target"].values <= prediction[:, :, "target_upper"].values)